requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pusher==3.3.2
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
